Refactor Auxiliary2AzureSearch to split steps into classes (#747)
Progress on https://github.com/NuGet/Engineering/issues/2978
