+++ src/NuGet.Services.AzureSearch/NuGet.Services.AzureSearch.csproj
@@ -49,6 +49,8 @@
     <Compile Include="Analysis\PackageIdCustomAnalyzer.cs" />
     <Compile Include="Analysis\PackageIdCustomTokenizer.cs" />
     <Compile Include="Analysis\TruncateCustomTokenFilter.cs" />
+    <Compile Include="Auxiliary2AzureSearch\UpdateDownloadsCommand.cs" />
+    <Compile Include="Auxiliary2AzureSearch\UpdateVerifiedPackagesCommand.cs" />
     <Compile Include="Auxiliary2AzureSearch\Auxiliary2AzureSearchCommand.cs" />
     <Compile Include="Auxiliary2AzureSearch\Auxiliary2AzureSearchConfiguration.cs" />
     <Compile Include="Auxiliary2AzureSearch\DownloadSetComparer.cs" />
